+++ ReleaseNotes.md
@@ -2,6 +2,11 @@
 
 # Release portal-pathogen
 
-## Release 1.4.0
+## Release 1.4.2
+- Updated ospo-resources for adding additional notes and disclaimer
 
+## Release 1.4.1
+- Add new font and background color
+
+## Release 1.4.0
 - First official GitHub-Release